Notification update

+++ backend/controllers/notification.controller.js
@@ -33,7 +33,34 @@ export const deleteNotifications = async (req, res) => {
 
     try {
         const currentUser = req.user._id
+        const notification =  await Notification.find({to:currentUser})
+        console.log(notification)
+        if (notification.length === 0) {
+            return res.status(404).json({ error: "Notification not found" })
+        }
          await Notification.deleteMany({to:currentUser})
+        return res.status(201).json({message:"Notifications deleted succesfully"})
+    } catch (error) {
+        console.log("Error in deleteNotifications controller", error.message)
+        return res.status(500).json({ error: "Internal Server Error in deleteNotifications controller" })
+    }
+};
+
+export const deleteOneNotifications = async (req, res) => {
+
+    try {
+        const {id} = req.params
+        const currentUser = req.user._id
+        const notification =  await Notification.findById({_id:id})
+        
+        if (!notification) {
+            return res.status(404).json({ error: "Notification not found" })
+        }
+
+        if (notification.to.toString() !== currentUser.toString()) {
+            return res.status(404).json({ error: "You are not allowed to delete this notification" })
+        }
+        await Notification.findByIdAndDelete(id)
         return res.status(201).json({message:"Notifications deleted succesfully"})
     } catch (error) {
         console.log("Error in deleteNotifications controller", error.message)
